+++ packages.css
@@ -1,11 +1,72 @@
-.card {
-display: flex;
-flex-wrap: wrap;
-justify-content: center;
-}
-#enterPackage{
-    height: 30px;
-}
-#enterDescription{
-    height: 150px;
-}
+body {
+    margin: 0.5em;
+  }
+  
+  .navbar {
+    margin-bottom: 2em;
+  }
+  .page-container {
+    display: flex;
+  }
+  
+  .user-container {
+    min-width: 20em;
+    min-height: 50em;
+    border: 1px solid black;
+    margin: 0 3em 0 3em;
+  }
+  
+  .card-container {
+    border: 1px solid black;
+    min-width: 20em;
+    min-height: 30em;
+    margin-bottom: 3em;
+    padding: 1em;
+    display: flex;
+    flex-wrap: wrap;
+  }
+  
+  .card-container div {
+      flex-basis: 33%;
+  }
+
+
+  .form-container {
+    display: flex;
+    align-content: flex-start;
+    justify-content: flex-start;
+    width: 77em;
+    align-items: center;
+    flex-direction: column;
+    border: 1px solid black;
+    padding: 5em;
+  }
+  
+  .form-container form {
+    width: 18em;
+    margin: 0.5em;
+  }
+  
+  .footer-container {
+    min-width: 70em;
+    min-height: 5em;
+    border: 1px solid black;
+    margin: 4em 3em 0 3em;
+  }
+  
+  .added-repo {
+    margin-left: 1em;
+    max-width: 50%;
+    text-align: justify;
+    text-justify: inter-word;
+  }
+  
+  .repo-list-item {
+    border-bottom: 1px solid gray;
+    margin: 0.5em;
+  }
+  
+  .list-group {
+    list-style-type: none;
+  }
+
